Updated QueryTest
In some cases, the consensus is just disordered, while the actual result is correct.
Let's perform a canonicalized assert in these cases.
There might be still some false positives (e.g. multidimensional comparisons), but that's okay, I guess.
Maybe, we can also find a way around that in the future.

+ Fixed empty string tokens ([''] and [""]) in Lexer

Resolves #9
Resolves #11
Resolves #12
Resolves #13
Resolves #14
Resolves #16
Resolves #18
Resolves #19
Resolves #21