Change getData() so that it can be mixed instead of array
The constructor itself allows $data to be mixed and the functions in
AccessHelper has various conditional statements to process the data
depending on whether it is an array or an implementation of ArrayAccess.
However, getData() is explicitly typed to return an array which means
any attempts to get the source object after creating a JSONPath of it
will cause a TypeError to be raised. Changing getData() to return mixed
instead of array will help fix this inconsistency.

Signed-off-by: Remy Suen <remy.suen@gmail.com>